+++ src/app/slice/mainSlice.jsx
@@ -29,7 +29,12 @@ export const get_quote = createAsyncThunk('/quote', async (payload, thunkAPI) =>
 
 export const get_studing = createAsyncThunk('/studing', async (payload, thunkAPI) => {
     try {
-        const { data } = await axios.get('http://localhost:3001/studing');
+        const { data } = await axios.get(process.env.REACT_APP_SERVER_URL + '/studing', {
+            headers: {
+                Authorization: `Bearer ${localStorage.token}`,
+            },
+        });
+        console.log(data);
         return thunkAPI.fulfillWithValue(data);
     } catch (error) {
         return thunkAPI.rejectWithValue(error);
