+++ src/css/timeTimer.module.css
@@ -60,3 +60,95 @@
     top: 11rem;
     font-size: 0.8rem;
 }
+
+.settingBtn {
+    width: 20.5rem;
+    height: 3.75rem;
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    align-items: center;
+    gap: 10px;
+    margin: 3.438rem 1rem 1rem;
+    padding: 0.75rem 1rem;
+    border-radius: 8px;
+    border: none;
+    background-color: var(--neutral-10);
+}
+
+.text {
+    width: 3.5rem;
+    height: 1.5rem;
+    color: var(--neutral-100);
+    font-size: 1rem;
+    font-weight: 600;
+    font-stretch: normal;
+    font-style: normal;
+    line-height: 1.5;
+    letter-spacing: normal;
+    text-align: center;
+}
+
+.setting {
+    width: 1.14rem;
+    height: 1.188rem;
+}
+
+.settingBox {
+    width: 1.5rem;
+    height: 1.5rem;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+.playBtn {
+    width: 20.5rem;
+    height: 3.75rem;
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    align-items: center;
+    gap: 10px;
+    margin: 3.438rem 1rem 1rem;
+    padding: 0.75rem 1rem;
+    border-radius: 8px;
+    border: none;
+    background-color: var(--primary-50);
+}
+
+.playBox {
+    width: 1.5rem;
+    height: 1.5rem;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+.playText {
+    width: 3.5rem;
+    height: 1.3rem;
+    color: var(--neutral-10);
+    font-size: 1rem;
+    font-weight: 600;
+    font-stretch: normal;
+    font-style: normal;
+    line-height: 1.5;
+    letter-spacing: normal;
+    text-align: center;
+}
+
+.mainTimerTime {
+    width: 11.5rem;
+    height: 3.938rem;
+    flex-grow: 0;
+    font-family: Pretendard;
+    font-size: 2.625rem;
+    font-weight: 600;
+    font-stretch: normal;
+    font-style: normal;
+    line-height: 1.5;
+    letter-spacing: normal;
+    text-align: center;
+    color: var(--neutral-100);
+}
